+++ unnamed_engine/src/core/event/window_event.rs
@@ -1,0 +1,7 @@
+use strum::Display;
+
+/// Events produced by a `winit::window::Window`.
+#[derive(Debug, PartialEq, Eq, Display)]
+pub enum WindowEvent {
+
+}